Do NOT give a ready-to-copy-paste /timezone command on detect-timezone.html
If you copy-paste the whole command, Discord is not going to register it as a command unless you edit it, click off the field or something like that. If you send it without doing that weird stuff to make Discord realize it's a command, it just gets sent as a regular message. So people should rather type /timezone themselves, then copy-paste the timezone. I swear Discord slash commands are drunk :a: